requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
qrcode
pillow
python-escpos
tzdata
requests
opencv-python
pyttsx3
python-multipart